requirements.txt
streamlit
google-cloud-firestore
pandas
numpy
plotly.express
datetime
streamlit-echarts
request
json